+++ Project Folder/DataLoadingAndPrep.py
@@ -1,0 +1,43 @@
+import os
+import numpy as np
+import matplotlib.pyplot as plt
+import torch
+from sklearn.datasets import load_digits
+from sklearn import datasets
+from torch.utils.data import Dataset, DataLoader
+import torch.nn as nn
+import torch.nn.functional as F
+import torchvision.transforms as tt
+
+
+
+class Digits(Dataset):
+    """Scikit-Learn Digits dataset."""
+
+    def __init__(self, mode='train', transforms=None):
+        digits = load_digits()
+        if mode == 'train':
+            self.data = digits.data[:1000].astype(np.float32)
+        elif mode == 'val':
+            self.data = digits.data[1000:1350].astype(np.float32)
+        else:
+            self.data = digits.data[1350:].astype(np.float32)
+
+        self.transforms = transforms
+
+    def __len__(self):
+        return len(self.data)
+
+    def __getitem__(self, idx):
+        sample = self.data[idx]
+        if self.transforms:
+            sample = self.transforms(sample)
+        return sample
+
+
+transforms = tt.Lambda(lambda x: 2. * (x / 17.) - 1.)
+train_data = Digits(mode='train', transforms=transforms)
+val_data = Digits(mode='val', transforms=transforms)
+test_data = Digits(mode='test', transforms=transforms)
+
+
